fix: refactor package names in internal/db directory and add a results group
