+++ internal/http/db/integration.go
@@ -1,4 +1,4 @@
-package integration
+package db
 
 import (
 	"hypha/api/internal/db"
