+++ internal/http/http.go
@@ -2,8 +2,7 @@ package http
 
 import (
 	"hypha/api/internal/db"
-	"hypha/api/internal/http/db/integration"
-	"hypha/api/internal/http/db/product"
+	db_group "hypha/api/internal/http/db"
 	"hypha/api/internal/http/report"
 	"hypha/api/internal/http/results"
 	"hypha/api/internal/utils/logging"
@@ -17,9 +16,11 @@ func InitRoutes(router *gin.Engine, dbOperations db.DatabaseOperations) {
 	log.Info().Msg("Initializing routes")
 
 	dbGroup := router.Group("/db")
-	product.InitProductRoutes(dbGroup, dbOperations)
-	integration.InitIntegrationRoutes(dbGroup, dbOperations)
-	results.InitResultsRoutes(dbGroup, dbOperations)
+	db_group.InitProductRoutes(dbGroup, dbOperations)
+	db_group.InitIntegrationRoutes(dbGroup, dbOperations)
+
+	resultsGroup := router.Group("/results")
+	results.InitResultsRoutes(resultsGroup, dbOperations)
 
 	reportGroup := router.Group("/report")
 	report.InitReportRoutes(reportGroup, dbOperations)
