+++ internal/http/results/utils.go
